Custom distribution buckets for pod e2e startup time metrics.
Specify custom bucket boundaries to extend the range of these metrics
up to 1 hour while keeping the number of buckets to a minimum and
preserving exact boundaries to support existing benchmarks and SLOs,
including "99% of pods with pre-pulled images start within 5 seconds".

The motivation for extending the range is to measure end-to-end pod
startup time in production clusters, where startup times can be much
longer than under test conditions.