Add deep links to PerfDash.
This change adds two-way binding between the drop-down menus of PerfDash
and the set of URL query parameters. The result is that changing the
drop-downs will change the URL, and changing the URL will update the UI.

More importantly, these URLs can be shared, causing PerfDash to load a
specific graph.