Prometheus - disable creating Kublet Endpoints
We don't monitor kubelets via Prometheus in scale tests, thus we can disable creation
of the Endpoints object for kubelets.

An alternative would be to wait for the Kubelet Endpoints object to be
created before running tests, but for now disabling the Kubelet Endpoints
object creation is a better solution.

Ref. https://github.com/kubernetes/kubernetes/issues/75284