Retry 404 errors when creating prometheus objects.
The 404 errors during object creations may be caused by races.
E.g. when a ServiceAccount is created the associated token is created asynchronously, when creating an object that needs the token we may get retryable 404 error if the token has not been created yet.

This should fix the issue described in https://github.com/kubernetes/kubernetes/issues/74213#issuecomment-476247400