Experimental Prometheus remote write to Cortex.
Inspired by the Cortex talks on KubeCon, I allowed myself to setup an experimental Cortex cluster.
The cluster is hosted in my dev project, the grafana for it is available at http://35.237.38.211:3000.

The advantages of using Cortex in addition to Prometheus are:
  * It supports multi-tenancy (we can have a separate tenant per test)
  * Horizontally scalable / highly available - it should be easily able to support the load from our tests (if what authors of Cortex are promising is true)
  * Faster / better query enging - we wouldn't run into prometheus "too many samples" limitations anymore

I'd like to experiment with it in our scalability tests.
If anything goes wrong it shouldn't break anything, given how the remote write is implemented in prometheus.

Ref. https://github.com/kubernetes/perf-tests/issues/523