Remove cpuInterval param from GetOneTimeResourceUsageOnNode()
It's not used anywhere other than to check that it's within some limits.
Also update the method doc, as it doesn't make sense comparing to what
the method actually does and how cadvisor works.

From the tests I've run, the cadvisor endpoint gets updated every ~15s
and scraping it takes on average ~110ms. It seems to be exactly the same
data that we scrape via prometheus. Given that, we should be able to
scrape it more often (in prometheus we scrape master every 5s) which
should allow us to fix
https://github.com/kubernetes/perf-tests/issues/694. This is a prework
to that.