Gather resource usage on master every 10s.
This is to address https://github.com/kubernetes/perf-tests/issues/694.

It's enough to scrape more often only on master as when it comes to nodes'
components we scrape them much more often on average, e.g. for 100 node
cluster we have 100 times more data for nodes components as each
node is scrape separately every 60s.

In addition it should be safe to hit cadvisor every 10s on master due to two
reasons:

1. We do it already via prometheus every 5s
2. From the tests I run looks like cadvisor updates every ~10-16s, so it
implements caching and requests are cheap (~100ms) and there is no gain
in scraping it often than 10s