Update cpuConstraints for scheduler and apiserver
Now, the https://github.com/kubernetes/perf-tests/pull/697 has merged,
the CPU usage is much less flaky. On the other hand it went up which is
expected as now we more accurately measure the actual 99th percentile
(see
https://github.com/kubernetes/kubernetes/issues/80212#issuecomment-515010726
as how bad we used to did it).

Given that we need to increase some constraints and remove the old
TODOs.