Extended Load: NetworkPolicies
This is most likely no-op until we turn on some Network Policy Provider that will start enfocring these network policies.

It should be pretty straightforward to turn on Calico both in GKE and in GCE.
This should be done separately to isolate any potential performance
impact of tuning it just on.

Ref. https://github.com/kubernetes/perf-tests/issues/704