Extend load test to cover PersistentVolumes
The only tricky part here is deleting PVs that are created via
StatefulSets. Theses PVs are not automatically deleted when StatefulSets
are deleted. Becasue of that I extended the ClusterLoader Phase api to
allow deleting objects that weren't created directly via CL2. The way
it works is that once we detect a new object if certain option is set we
issue a List request to find nunmber of replicas.

Ref. https://github.com/kubernetes/perf-tests/issues/704