Add storage experimental test suite
This adds a test suite for running storage tests. In the next commit (to
test-infra), a prow job that runs this suite will be defined in
sig-scalability-experimental-periodic-jobs.

The goal of running a suite instead of duplicated jobs is to minimize
config duplication.

To keep metric file names simple in case of a test suite, we use empty
test name override. This will be removed once migration from separate
storage test runs to a single suite run is done.

There is some redundancy in the suite definition. To be investigated if
it can be reduced using go templates. If it is feasible, it will be done
as a follow-up change once the suite job is working.

Tesing:
* ran clusterloader with
testsuite=testing/experimental/storage/pod-startup/suite.yaml and
verified that tests passed