Add support for displaying test suite results in perfdash
This commit enables perfdash to view metrics recorded by clusterloader
test suite runs.

Before this change, it was not possible to do so, because test suite run
puts measurement results in separate files, for example:
-  APIResponsiveness_storage_configmap-vol-per-node_2019-10-01T06:08:05Z.json
-  APIResponsiveness_storage_downwardapi-vol-per-node_2019-10-01T06:10:00Z.json
In such case, perfdash read metrics data from the first file that
matched the prefix APIResponsiveness_storage (prefix is specified in
config.go file), meaning that other results were ignored.

To solve this issue, support for handling multiple files that match
given prefix is added. For each of the matched files, perfdash extracts
the "suiteId" from the file name and appends it to the result category.
This means that results from different test cases in a given suite will
land in different categories in the web UI, for example:
- APIServer_configmap-vol-per-node
- APIServer_downwardapi-vol-per-node

Alternatively, same effect could be achieved by generating separate test
descriptions in the perfdash config. It was not chosen, because it would
mean that perfdash config needs to be changed every time some test case
is added or removed.

This commit also includes other minor, but related changes like:
- increasing with of category field in the web UI
- removing 'pod-with-ephemeral-volume' configs

Testing:
- run perfdash locally, verify that data for both storage-suite and
separate storage jobs is displayed as expected