Switch Prometheus API responsiveness to use 1m windows.
Default so far was 5m as in SLI definition. We're changing this to 1m, so we
have we calculate the the SLI over more data points. In the past, we were
suffering from having not enough of them, which was causing 99th percentaile to
be actually max over the series.

Minor: remove recording rules with 30s window.