Add system pod metrics measurement in clusterloader
In this commit, "system pod metrics" measurement is added. Its purpose
is to gather data about system pod restarts. This measurement will be
invoked on every test run, but it's logic is no-op by default.

It will be enabled by setting "systemPodMetricsEnabled: true" in
TestMetrics measurement. This is to be done in follow-up commits.

Testing:
* ran storage suite against existing cluster
* verified that SystemPodMetrics artifact is being created
with proper content if flag is set to true
* verified that no such artifact is created when this flag is absent