Subtract initial restart counts when calculating system pod metrics
Before this change, system pod measurement gathered metrics only at the
end of each test. This means that the restart count in the present in
the summary was equal to the restart count since each given container
was created (this is usually during cluster creation). This means that
we could have high restart count metric for some test even though no
restart happened during that test, which could be misleading.

To fix this, in this commit we add listing system pods before the test
(in measurement "start" action). Then in "gather" action, initial
restart counts are subtracted from the actual restart counts. In effect,
restart count present in the summary is the number of restarts that
occured during execution of this test.

Testing:
* ran storage suite locally
* verified SystemPodMetrics summary is being created
* verified that test fails if we execute gather without executing start
action first
* unit tests for the restart count subtraction logic