Enable system pod metrics collection by default
This will enable collection of system pod metrics by default in all
tests. In previous step, it was enabled in gce-5000 test and metrics
were successfully gathered (with measurement time less than 20s).

In the follow-up commits, overrides in test-infra will be removed.
However, this flag will be still in place for some time to be able to
turn this feature off easily if there are any issues.