Allow changing default allowed restart count
This feature is mostly used for detecting crashlooping system
components. Single restart of any component may be ok, see discussion in
https://github.com/kubernetes/perf-tests/pull/908.

At this point, tests will fail if any container (which is not present in
overrides) has even one restart. To simplify this, in this change
support for 'default' max allowed restart count was added. Setting it
will be as simple as adding 'default: X' in the overrides map. Note that
this won't be applied for containers that have some threshold specified.