Implement fetching prow config from directories
At this point, we need to specify each prow config file explicitly in
deployment yaml. This will be a problem for adding sig-release jobs to
pefdash, since these configs are autogenerated by config forker.

With this change, we'd be able to just specify "githubConfigDir" to the
sig-release dir and perfdash will ingest all the configs from that dir.

Testing:
* started perfdash locally with
githubConfigDir=path-to-sig-release-configs, verified configs are
fetched
* tested that perfadash still works without providing githubConfigDir
flag