Simplify logging in the isNodeConditionSetAsExpected
It's an antipattern to control level of logging via method arguments, a
logging framework features should be used for that.
Changing the log level to V(4) so it can be turned on/off by setting the
--v flag appropriately.

A justfication for making this change is the CA tests where nodes appear
and disappear, in the current for this logging spams the CL2 output.