port number of etcd insecure access should be configurable
The port number should be configurable instead of hard-coded,
actually, for the cluster which is setup by kubeadm, the insecure
access number is "2381" by default.

It's hard-coded as "2382", this will make end-user faces issues
when etcd metrics is needed, change source code to gather data for
end-user does not make sense, we'd better to provide a option so that
the port number could be specified before the job is started.