clean up stale namespaces before execution
It's not likely execution is finished normally everytime, the execution
might be aborted abnormally, this will make the auto managed namespaces
become orphans and no chance to be clean up on the test machines, thus,
impact the data we got from execution.

The change gives a chance to clean up all those namespaces before the
execution, make the execution a fairly clean environment.