feat[Clusterloader2][NodeKiller]: wait for NodeKiller to restore nodes
With this change I tried to introduce some sort of wait mechanism for the NodeKiller subroutine.
The rationale behind is it avoiding that cluster nodes are left in a broken state by NodeKiller if the test finishes while downtime is being simulated for some nodes.

Now the chaos monkey can return back to its caller a WaitGroup that can be used to wait for the subroutine termination before proceeding to test summary.

Fixes #1130
Signed-off-by: Manuel Castellin <manuel@castellinconsulting.com>