Unpack the string values from the CL2_.* envs
Without doing it, it's impossible to easily override boolean param via env. E.g. if I do
```
--env=ENABLE_MY_FEATURE=false
```

The `false`  will be treated as string and then in the config file the if statement like this:
```
{{if .ENABLE_MY_FEATURE}}
...
{{end}}
```

Will evaluate to true and will be executed, because "false" is non-empty string that evaluates to true :)