Collect resource usage from non-joined master nodes over SSH
Some providers do not have master nodes joined to a cluster, so
fallback to SSH is added for such providers.

Resource gatherer will merge nodes registered in a cluster with
master ips, deduplicating by address, and will use old method of
collecting container ids and gathering metrics, when possible.

For master ips that are not joined to a cluster as a node, gatherer
will use provider's SSH method, and will collect metrics of all
containers registered on that node, including ones running
as part of static pods.

Kubelet package is extended with ssh stats gatherer, and generalized
way to provide container filtering (either list based, or match-all).