cl2/density: add pod topology spread workload to scheduler suite
This pod topology spread workload uses the hostname as the topology key
create a spreading deployment with a max skew of 1. This is effectively
a round-robin.

The hostname was used as the topology key due to the lack of a better
label. Ideally, I would've wanted to benchmark zone spreading with
topology.kubernetes.io/zone as the topology key, but since all Kubemark
kubemark nodes have identical labels except for kubernetes.io/hostname,
this is the only viable option. In the future, we can look into the
following options to make this more even representative of a real
workload:

1. Make Kubemark hollow nodes be actually present in different GCP
   zones, thereby guaranteeing that each node has a different
   topology.kubernetes.io/zone.
2. Set a dummy label on hollow nodes to be used as the spreading
   topology key at Kubemark cluster creation time. This would probably
   be a kubetest change.
3. Allow setting and clearing node labels as a part of clusterloader.

I think (3) is the cleanest option if we end up doing this, but whether
or not setting node labels is within the scope of clusterloader is
debatable.

Signed-off-by: Adhityaa Chandrasekar <adtac@google.com>