pod-startup: flexible volume provisioning
For basic sanity testing it is fine to use a minimal number of
PODS_PER_NODE and VOLUMES_PER_POD. The storage class name shouldn't be
hardcoded to csi-gce-pd.

There is nothing specific to CSI about it, therefore the USE_CSI
parameter can be removed.

Metrics gathering can be slow. When the goal is to just verify that
volume provisioning works correctly under load, then disabling metrics
gathering is useful.

When investigating how quickly volumes can be provisioned, avoiding
the creation of pods is useful.

One observation is that provisioning gets slowed by the PV controller
which needs to add the "provisioner" annotation. That part
can (optionally) be skipped by creating PVCs which already have that
annotation. In that case, not waiting for volume binding which again
is done by the PV controller is also useful.

Other changes:

- allow overriding timeout

The expected duration depends greatly on the volume size and CSI
driver and thus should be configurable.

- PV utils: fix cut-and-paste error in error message

When addedPVs is non-empty, we want to dump that and not deletedPVs...

- PV utils: include one example PV in error message

During scale testing, one PV refused to go away. The extra output
shows what clusterload2 was waiting for (a PV for Prometheus which
wasn't created by the test).

- support running on a cluster with existing PVs

There are currently no labels that could be used to filter for PVs
that were created by clusterloader2. What we can use is the
"provisioned-by" annotation. That's still not perfect (there might be
PVs created by the provisioner under test), but better than nothing.