chore: fork the metricsGrabber code into pref-tests repo
The upstream is going to disable the insecure port of controller-manager
and scheduler, and now this is being blocked by `clusterloader2` because
it vendor test/e2e from k/k v1.18 and grab metrics from the insecure port
in a test job.

This commit copies the metricsGrabber code into perf-tests repo and make
metricsGrabber grab metrics from the secure port of both
controller-manager and scheduler.