set firewall port to netperf data path and set IP version for iperf3
- the hairpin test with iperf3 was failing in certain kernel versions
(3.10) because the server was opening a IPv6 socket instead of IPv4
and the connection was reset during test. The same problem was not
observed on later version (4.18). The change makes an explicit protocol
version set for the iperf3 server, based on pod-IP version
- tests with netperf against virtual-IP need to add two TCP ports, one for
control and another for the data path, otherwise the test will be blocked
by the firewall rules. The change add a port on the firewall and passes it
to the netperf client