Fix the port number to collect the profiling data for the apiserver
`443` is the tcp port for the cluster ip `10.96.0.1`, which is used by
service `kubernetes`.

Connect the port with the localhost and fetch the profiling data is not allowed,

> curl -k https://localhost:443/debug/pprof/profile?seconds=30 -o cpu2.pprof

  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0
curl: (7) Failed to connect to localhost port 443: Connection refused

Because of this, profiling file for apiserver is always empty,

> file 10.253.2.51_kube-apiserver_CPUProfile_load_2021-08-26T11:10:26Z.pprof
10.253.2.51_kube-apiserver_CPUProfile_load_2021-08-26T11:10:26Z.pprof: empty

We should either update the url to use clusterip:443 or just update the port
to use the port exposed by the pod directly.

Signed-off-by: Dave Chen <dave.chen@arm.com>