Add separate huge service test config
We want to decouple scheduler throughput from huge service test. In
order to do so, a separate test needs to be implemented. Since we do not
test for huge services everywhere, it is done in a separate file.

The default number of endpoints for the new huge service test is 1000
instead of the number of nodes.