clusterloader2 pod startup: wait for pod deletion
The test needs to wait for deletion, at the latest when cleaning up.
Having an explicit step makes it more obvious what it is waiting for, in
particular because that then prints regular status updates.

When using generic ephemeral inline volumes and waiting for volume
deletion (another explicit step), not waiting for pod deletion first gave the
wrong impression that volume deletion is slow. What really happened was that
volume deletion kept up with pod deletion and only that was slow.