Run command from the first word of the first arg in run-e2e.sh
run-e2e.sh is usually ran from kubetest (test-infra). It is defined by
repeatedly declaring test-cmd-args argument, which is itself
a leftover after using pflag dependency. It may be wanted to use only
a single declaration of this argument. In such a case, the first
argument will contain all the parameters for ClusterLoader2.

This change makes the approach above possible by setting the tool name
to the first word of the first argument value and adds the remainder
as additional flags for the clusterloader2/run-e2e.sh command.