clusterloader2: use new apiserver latency SLI
In Kubernetes 1.23.0 a new request latency metric more suitable for SLOs
since it doesn't account for webhooks latency was introduced. Since this
new metric is better suited to clusterloader2 than the general purpose
one, it would make sense to update the existing Prometheus rules.

Note that this implies moving from a STABLE metric to an ALPHA one, but
considering that the new metric has been around for already 3 releases
and that it has a real-life scenario attached to it, the metric is very
unlikely to be removed in the future.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>