clusterloader2: rules and dashboards version aware
The apiserver measurement recording rule and the various dashboard were
all using the apiserver_request_duration_seconds metrics, however we
want to move to the new slo/sli metrics since they give more precise
latency measurements. Because these metrics are only available in
certain versions of Kubernetes, we needed to make the recording rules
and dashboards version aware.

For the recording rules used for measure, we want the metrics that we
use to be as precise as possible so we chose to deduplicate the existing
queries with the new metrics and select which one to use in code
depending on the cluster version.

For dashboards it is a bit more complex since we can't change the query
at runtime, so instead of taking the approach that we took for
measurement, we just added a new recording rule that would use the
correct metric depending on the cluster version. This as the benefit of
not being intrusive with the existing code, but at the same time since
the result are now precomputed, the data will be less fresh than it used
to be, but the difference is not very significant from a dashboard
perspective.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>