cl2: Add support for fatal errors during measurement execution
Currently, CL2 measurements are unable to cause the entire test to bail
if they fail to execute. The errors returned from a measurements
`Execute()` method are passed through a function named
`isErrsCritical()` by the simple test executor, however this function is
a no-op. This commit introduces a new error type, `ErrFatal`, which
when returned by a measurement, causes the simple test executor to bail.

An example use case is demonstrated in the following commit.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>