cl2: Add `isFatal` param to WaitForRunningPods measurement
This commit utilizes the previous commit to add a field named `isFatal`
to the WaitForRunningPods measurement. When set to `true`, and the
measurement is unable to verify that the targeted pods are ready before
the timeout expires, a fatal error is propagated to the simple test
executor which causes the entire test to fail.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>